Validate user inputs and ensure code base is documented for better description.
