+++ server/controllers/books.js
@@ -1,24 +1,41 @@
 import model from '../models';
-import helpers from '../helpers';
 
 const Book = model.Book;
-const isAdmin = helpers.isAdmin;
-const validateInput = helpers.validateInput;
 
 /**
  * 
  */
 class BookController {
   /**
-   *
-   * @param {*} req
-   * @param {*} res
+   * 
+   * 
+   * @static
+   * @param {any} req 
+   * @param {any} res 
+   * @returns 
+   * 
+   * @memberOf BookController
    */
   static create(req, res) {
-    // Check if user has administrative priviledges
-    if (!isAdmin(req)) {
+    // Ensure user has administrative priviledges to create book
+    if (req.decoded.data.role !== 'admin') {
       return res.status(403).send({ success: false, message: 'Permission Denied' });
     }
+    // Validates every input by the user.
+    if (Object.keys(req.body).length < 5) return res.status(400).send({ success: false, message: 'All fields are required.' });
+    const errors = [];
+    for (let i = 0; i < 5; i += 1) {
+      if (Object.values(req.body)[i] === (undefined || null || '')) {
+        errors.push(`${Object.keys(req.body)[i]} field is required`);
+      }
+      if (Object.keys(req.body)[i] === 'quantity' && typeof (parseInt(Object.values(req.body)[i], 10)) !== 'number') {
+        errors.push('quantity must be a number');
+      }
+    }
+    // Error(s) is/are outputted if any is pushed to the array
+    if (errors.length > 0) return res.status(400).send(errors);
+
+    // Searches if book exists in the database
     return Book.findOne({
       where: { isbn: req.body.isbn },
     })
@@ -26,6 +43,7 @@ class BookController {
         if (foundBook) {
           return res.status(409).send({ success: false, messsage: `Conflict! ${req.body.title} exists already`, foundBook });
         }
+        // If book does not exist, create new book.
         return Book
           .create({
             isbn: req.body.isbn,
@@ -39,19 +57,40 @@ class BookController {
           .then((book) => {
             res.status(200).send({ success: true, message: `${book.title}, succesfully added` });
           })
-          .catch(() => { validateInput(req, res); });
+          .catch(error => res.send(error.message));
       })
-      .catch(() => { validateInput(req, res); });
+      .catch(error => res.send(error.message));
   }
   /**
-   *
-   * @param {*} req
-   * @param {*} res
+   * 
+   * 
+   * @static
+   * @param {any} req 
+   * @param {any} res 
+   * @returns 
+   * 
+   * @memberOf BookController
    */
   static update(req, res) {
-    if (!isAdmin(req)) {
+    // Ensure user has administrative priviledges to create book
+    if (req.decoded.data.role !== 'admin') {
       return res.status(403).send({ success: false, message: 'Permission Denied' });
     }
+    // Validates every input by the user.
+    if (Object.keys(req.body).length < 5) return res.status(400).send({ success: false, message: 'All fields are required.' });
+    const errors = [];
+    for (let i = 0; i < 5; i += 1) {
+      if (Object.values(req.body)[i] === (undefined || null || '')) {
+        errors.push(`${Object.keys(req.body)[i]} field is required`);
+      }
+      if (Object.keys(req.body)[i] === 'quantity' && typeof (parseInt(Object.values(req.body)[i], 10)) !== 'number') {
+        errors.push('quantity must be a number');
+      }
+    }
+    // Error(s) is/are outputted if any is pushed to the array
+    if (errors.length > 0) return res.status(400).send(errors);
+
+    // Checks if book exists in the database
     return Book
       .findOne({
         where: {
@@ -62,6 +101,7 @@ class BookController {
         if (!book) {
           return res.status(404).send({ success: false, message: 'Book not found!' });
         }
+        // If book exists, update the book.
         return Book
           .update({
             isbn: req.body.isbn,
@@ -82,20 +122,25 @@ class BookController {
             }
             return res.status(200).send({ success: true, message: `${req.body.title}, successfully updated`, old: book });
           })
-          .catch(() => {
-            validateInput(req, res);
-          });
+          .catch(error => res.send(error.message));
       });
   }
   /**
    * 
-   * @param {*} req 
-   * @param {*} res 
+   * 
+   * @static
+   * @param {any} req 
+   * @param {any} res 
+   * @returns 
+   * 
+   * @memberOf BookController
    */
   static destroy(req, res) {
+    // Ensures user has administrative priviledges to delete book
     if (req.decoded.data.role !== 'admin') {
-      return res.status(403).send({ success: false, message: 'You are not allowed to delete books' });
+      return res.status(403).send({ success: false, message: 'Permission Denied' });
     }
+    // Searches for book in the database
     return Book
       .find({
         where: {
@@ -106,6 +151,7 @@ class BookController {
         if (!book) {
           return res.status(404).send({ success: false, message: 'Book not found' });
         }
+        // If book is found, delete
         return book.destroy();
       })
       .then(() => {
@@ -113,9 +159,14 @@ class BookController {
       }).catch(() => res.status(400).send({ success: false, message: 'Enter valid inputs!' }));
   }
   /**
-   *
-   * @param {*} req
-   * @param {*} res
+   * 
+   * 
+   * @static
+   * @param {any} req 
+   * @param {any} res 
+   * @returns 
+   * 
+   * @memberOf BookController
    */
   static list(req, res) {
     return Book
