+++ server/controllers/borrows.js
@@ -8,10 +8,16 @@ const Borrow = model.Borrow;
 class BorrowController {
   /**
    * 
-   * @param {*} req 
-   * @param {*} res 
+   * 
+   * @static
+   * @param {any} req 
+   * @param {any} res 
+   * @returns 
+   * 
+   * @memberOf BorrowController
    */
   static create(req, res) {
+    // First checks if book has been borrowed and not returned
     return Borrow
       .findOne({
         where: {
@@ -23,9 +29,14 @@ class BorrowController {
           { model: Book, as: 'book', required: true },
         ],
       }).then((foundBorrow) => {
+        /**
+         * If book has been borrowed before and not returned,
+         * User cannot borrow same book again.
+         */
         if (foundBorrow) {
           return res.status(409).send({ success: false, messsage: 'Conflict! Book borrowed already', foundBorrow });
         }
+        // Else, user is eligible to borrow book
         return Borrow
           .create({
             returned: false,
@@ -35,6 +46,7 @@ class BorrowController {
             actualReturnDate: Date.now(),
           })
           .then(() => {
+            // Ensures book is available and not borrowed out.
             Book
               .findOne({
                 where: {
@@ -42,9 +54,14 @@ class BorrowController {
                 },
               })
               .then((foundBorrowedBook) => {
+                // If book is borrowed out, then No book to borrow
                 if (!foundBorrowedBook || foundBorrowedBook.quantity === 0) {
                   return res.status(404).send({ success: false, message: 'Book not found' });
                 }
+                /**
+                 * But if book is available, User can borrow book
+                 * with the count decreased by one
+                 */ 
                 return foundBorrowedBook
                   .update({
                     quantity: foundBorrowedBook.quantity - 1,
@@ -66,8 +83,18 @@ class BorrowController {
         res.status(400).send({ success: false, message: `drOops! something happened, ${error.message}` });
       });
   }
-
+  /**
+   * 
+   * 
+   * @static
+   * @param {any} req 
+   * @param {any} res 
+   * @returns 
+   * 
+   * @memberOf BorrowController
+   */
   static returnBook(req, res) {
+    // Check if book has been borrowed before and returned
     return Borrow
       .findOne({
         where: {
@@ -79,17 +106,20 @@ class BorrowController {
           { model: Book, as: 'book', required: true },
         ],
       }).then((foundBorrow) => {
+        // If borrowed and returned, then you have no business returning again.
         if (foundBorrow) {
           return res.status(409).send({ success: false, messsage: 'Conflict! Book returned already', foundBorrow });
         }
+        // Else, allows to return
         return Borrow
           .update({
-            returned: true,
-            actualReturnDate: Date.now(),
+            returned: true, // Sets returned status to true
+            actualReturnDate: Date.now(), // Updates returned date to now
           }, {
             where: { userId: req.params.userId, bookId: req.body.bookId },
           })
           .then(() => {
+            // Searches if book is available in the database
             Book
               .findOne({
                 where: {
@@ -100,6 +130,7 @@ class BorrowController {
                 if (!foundBorrowedBook || foundBorrowedBook.quantity === 0) {
                   return res.status(404).send({ success: false, message: 'Book not found' });
                 }
+                // Book is returned with the count increased by one
                 return foundBorrowedBook
                   .update({
                     quantity: foundBorrowedBook.quantity + 1,
@@ -121,8 +152,18 @@ class BorrowController {
         res.status(400).send({ success: false, message: `Oops! something happened, ${error.message}` });
       });
   }
-
+  /**
+   * 
+   * 
+   * @static
+   * @param {any} req 
+   * @param {any} res 
+   * @returns 
+   * 
+   * @memberOf BorrowController
+   */
   static listNotReturned(req, res) {
+    // List all books borrowed but not returned by a User.
     return Borrow
       .findAll({
         where: {
